serviceproxy/azure_openai/requirements.txt
openai>=1.54.5,<2
python-dotenv>=1.0.1,<2
